🎨 api cleanups  - move nested converters to top level  - provide default constructors for converters that throw
