+++ kivakit-security/pom.xml
@@ -33,6 +33,11 @@
             <artifactId>kivakit-conversion</artifactId>
             <version>${project.version}</version>
         </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>kivakit-resource</artifactId>
+            <version>${project.version}</version>
+        </dependency>
 
         <!-- Annotations -->
 
@@ -49,12 +54,6 @@
 
         <!-- Test -->
 
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>kivakit-core</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-        </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>kivakit-testing</artifactId>
